+++ src/components/layout/YouthLayout.tsx
@@ -1,0 +1,58 @@
+import React, { useState } from 'react';
+import { YouthSidebar } from './YouthSidebar';
+import { Header } from './Header';
+import { Menu } from 'lucide-react';
+import { Button } from '../ui/Button';
+
+interface YouthLayoutProps {
+  children: React.ReactNode;
+}
+
+export const YouthLayout = ({ children }: YouthLayoutProps) => {
+  const [sidebarOpen, setSidebarOpen] = useState(false);
+
+  return (
+    <div className="flex h-full bg-[#F5F5F0] dark:bg-gray-900">
+      {/* Mobile sidebar overlay */}
+      {sidebarOpen && (
+        <div 
+          className="fixed inset-0 z-40 lg:hidden bg-black bg-opacity-50"
+          onClick={() => setSidebarOpen(false)}
+        />
+      )}
+      
+      {/* Desktop sidebar */}
+      <div className="hidden lg:flex lg:flex-shrink-0">
+        <YouthSidebar />
+      </div>
+      
+      {/* Mobile sidebar */}
+      <div className={`fixed inset-y-0 left-0 z-50 lg:hidden transform transition-transform duration-300 ease-in-out ${
+        sidebarOpen ? 'translate-x-0' : '-translate-x-full'
+      }`}>
+        <YouthSidebar onClose={() => setSidebarOpen(false)} />
+      </div>
+      
+      {/* Main content */}
+      <div className="flex-1 flex flex-col overflow-hidden">
+        {/* Header with mobile menu button */}
+        <div className="flex items-center justify-between p-4 bg-white dark:bg-gray-800 border-b border-[#B45309]/20 lg:hidden">
+          <Button
+            variant="ghost"
+            size="sm"
+            onClick={() => setSidebarOpen(true)}
+            className="lg:hidden"
+          >
+            <Menu className="h-5 w-5" />
+          </Button>
+          <Header />
+        </div>
+        
+        {/* Page content */}
+        <main className="flex-1 overflow-y-auto">
+          {children}
+        </main>
+      </div>
+    </div>
+  );
+};
