+++ src/components/layout/YouthSidebar.tsx
@@ -10,7 +10,8 @@ import {
   Folder, 
   X, 
   Calendar,
-  ChevronRight
+  ChevronRight,
+  Briefcase
 } from 'lucide-react';
 import { useAuth } from '../../utils/auth';
 import { Button } from '../ui/Button';
@@ -110,10 +111,10 @@ export const YouthSidebar = ({ onClose }: YouthSidebarProps) => {
         <ul className="space-y-1">
           <li>
             <Link 
-              to="/dashboard" 
+              to="/" 
               onClick={handleLinkClick} 
               className={`flex items-center px-4 py-3 rounded-md transition-all duration-200 ${
-                isActive('/youth-dashboard') 
+                isActive('/') && location.pathname === '/' 
                   ? 'bg-[#B45309]/10 text-[#B45309]' 
                   : 'text-[#503314] hover:bg-[#F5F5F0] dark:text-white dark:hover:bg-gray-700'
               } ${!shouldShowText ? 'justify-center' : ''}`}
@@ -121,6 +122,22 @@ export const YouthSidebar = ({ onClose }: YouthSidebarProps) => {
             >
               <Home className="h-5 w-5 flex-shrink-0" />
               {shouldShowText && <span className="ml-3">Dashboard</span>}
+            </Link>
+          </li>
+          
+          <li>
+            <Link 
+              to="/opportunities" 
+              onClick={handleLinkClick} 
+              className={`flex items-center px-4 py-3 rounded-md transition-all duration-200 ${
+                isActive('/opportunities') 
+                  ? 'bg-[#B45309]/10 text-[#B45309]' 
+                  : 'text-[#503314] hover:bg-[#F5F5F0] dark:text-white dark:hover:bg-gray-700'
+              } ${!shouldShowText ? 'justify-center' : ''}`}
+              title={!shouldShowText ? "Opportunities" : ""}
+            >
+              <Briefcase className="h-5 w-5 flex-shrink-0" />
+              {shouldShowText && <span className="ml-3">Opportunities</span>}
             </Link>
           </li>
