+++ src/pages/LandingPage.tsx
@@ -96,8 +96,8 @@ export const LandingPage = () => {
       </section>
 
       {/* About Us Section */}
-      <section id="aboutUs" className="py-20 bg-[#F5F5F0] dark:bg-gray-900 relative overflow-hidden">
-        <RwandaMap className="opacity-5 w-screen h-[55vh] sm:h-[60vh] 2xl:h-[70vh] absolute group-hover:scale-105 z-40"/>
+      <section id="aboutUs" className="py-20 bg-[#F5F5F0] dark:bg-gray-900 relative min-h-screen overflow-hidden">
+        <RwandaMap className="opacity-5 w-screen h-[60vh] md:h-[70vh] lg:h-[90vh]  absolute group-hover:scale-105 z-40"/>
         <div className="container mx-auto px-4 relative z-50">
           <div className="text-center mb-16 animate-fadeIn">
             <div className="inline-flex items-center px-4 py-2 bg-[#B45309]/10 rounded-full border border-[#B45309]/20 mb-6">
